Replace `node-cron` plugin with Nitro native `scheduledTasks`
