+++ dashboard/package.json
@@ -29,7 +29,6 @@
     "jszip": "^3.10.1",
     "kysely": "^0.28.17",
     "lodash-es": "^4.17.21",
-    "node-cron": "^4.2.1",
     "nuxt": "^4.3.1",
     "papaparse": "^5.5.3",
     "pinia": "^3.0.3",
@@ -45,7 +44,6 @@
     "@types/color": "^4.2.1",
     "@types/lodash-es": "^4.17.12",
     "@types/node": "^24.6.0",
-    "@types/node-cron": "^3.0.11",
     "@types/papaparse": "^5.5.1",
     "@vitejs/plugin-vue": "^6.0.1",
     "@vue/eslint-config-prettier": "^10.2.0",
