+++ dashboard/server/tasks/expire-sessions.ts
@@ -1,0 +1,9 @@
+export default defineTask({
+  meta: {
+    description: 'Delete expired user sessions.',
+  },
+  async run() {
+    await deleteExpiredSessions()
+    return { result: 'ok' }
+  },
+})
